[refactor] Toy, Feed 클래스 위치변경
